refactor : enum 클래스 공통된 enumerate 패키지로 이동
