+++ src/main/java/nextstep/courses/enumerate/SessionStatusType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.enumerate;
 
 public enum SessionStatusType {
     PREPARATION, RECRUITING, END
