+++ src/test/java/nextstep/courses/domain/CoverImageTest.java
@@ -3,6 +3,7 @@ package nextstep.courses.domain;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import nextstep.courses.CanNotCreateException;
+import nextstep.courses.enumerate.CoverImageType;
 import org.junit.jupiter.api.Test;
 
 class CoverImageTest {
